Alteração de interface Turma
-Alteração de interface no formulario da turma e inserção de titulo
-Na tela de listagem, implementação de paginação e sort por coluna, assim como botao de geração de relatorio.

+++ UI/TrackClassApp/src/app/features/turma/turma-list/turma-list.component.ts
@@ -1,9 +1,12 @@
 import { Component, OnInit } from '@angular/core';
 import { TurmaService } from '../../../core/services/turma.service';
-import { Turma } from '../../../core/models/turma.model';
 import { Router } from '@angular/router'
 import { TurmaAlunoService } from '../../../core/services/turma-aluno.service';
 import { MessageService } from 'primeng/api';
+import jsPDF from 'jspdf';
+import 'jspdf-autotable';
+import { ConfirmDialogComponent } from '../../confirm-dialog/confirm-dialog.component';
+import { MatDialog } from '@angular/material/dialog';
 
 @Component({
   selector: 'app-turma-list',
@@ -12,12 +15,16 @@ import { MessageService } from 'primeng/api';
   providers: [MessageService]
 })
 export class TurmaListComponent implements OnInit {
-  turmas: Turma[] = [];
-  filteredTurmas: Turma[] = [];
-  filter: string = '';
   displayedColumns: string[] = ['ativo', 'nome', 'descricao'];
+  sortField: string = '';
+  sortOrder: number = 1;
+  turmas: any[] = [];
+  filteredTurmas: any[] = [];
+  filter: string = '';
 
-  constructor(private turmaService: TurmaService,  private turmaAlunoService: TurmaAlunoService,private messageService: MessageService, private router: Router) { }
+  constructor(private turmaService: TurmaService,  private turmaAlunoService: TurmaAlunoService,
+    private messageService: MessageService, private router: Router,
+    public dialog: MatDialog) { }
 
   ngOnInit(): void {
     this.turmaService.getTurmas().subscribe(turmas => {
@@ -34,26 +41,33 @@ export class TurmaListComponent implements OnInit {
   }
 
   onDelete(id: number): void {
-    this.turmaAlunoService.existeTurmaVinculada(id).subscribe(
-      (exists) => {
-        if (exists) {
-          this.messageService.add({ severity: 'error', summary: 'Erro', detail: 'A turma ja esta vinculada a um aluno, favor desvincular' });
-        } else {
-          this.turmaService.deleteTurma(id).subscribe(
-            (response) => {
-              this.messageService.add({ severity: 'success', summary: 'Sucesso', detail: 'Turma excluida com sucesso' });
-              this.ngOnInit();
-            },
-            (error) => {
-              this.messageService.add({ severity: 'error', summary: 'Erro', detail: 'Erro ao excluir Turma' });
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      data: { message: 'Tem certeza que deseja excluir esta turma?' }
+    });
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.turmaAlunoService.existeTurmaVinculada(id).subscribe(
+          (exists) => {
+            if (exists) {
+              this.messageService.add({ severity: 'error', summary: 'Erro', detail: 'A turma ja esta vinculada a um aluno, favor desvincular' });
+            } else {
+              this.turmaService.deleteTurma(id).subscribe(
+                (response) => {
+                  this.messageService.add({ severity: 'success', summary: 'Sucesso', detail: 'Turma excluida com sucesso' });
+                  this.ngOnInit();
+                },
+                (error) => {
+                  this.messageService.add({ severity: 'error', summary: 'Erro', detail: 'Erro ao excluir Turma' });
+                }
+              );
             }
-          );
-        }
-      },
-      (error) => {
-        this.messageService.add({ severity: 'error', summary: 'Erro', detail: 'Erro ao verificar Turma ' });
+          },
+          (error) => {
+            this.messageService.add({ severity: 'error', summary: 'Erro', detail: 'Erro ao verificar Turma ' });
+          }
+        );
       }
-    );
+    });
   }
 
   onEdit(id: number): void {
@@ -66,5 +80,54 @@ export class TurmaListComponent implements OnInit {
 
   newTurma(){
     this.router.navigate(['/turma/new']);
+  }
+
+  gerarRelatorioPDF(): void {
+    const doc = new jsPDF();
+    let y = 15;
+
+    doc.setFontSize(18);
+    doc.text('Relatório de Turmas', 10, y);
+
+    const headers = [['Nome', 'Descrição', 'Ativo']];
+    const data = this.turmas.map(turma => [
+        turma.nome,
+        turma.descricao,
+        turma.ativo ? 'Sim' : 'Não'
+    ]);
+
+    (doc as any).autoTable({
+        startY: 25,
+        head: headers,
+        body: data,
+        theme: 'grid',
+        styles: { cellPadding: 3, fontSize: 12 },
+        columnStyles: { 4: { cellWidth: 'wrap' } }
+    });
+
+    doc.save('relatorio_turmas.pdf');
+  }
+
+  customSort(event: any): void {
+    this.sortField = event.field;
+    this.sortOrder = event.order;
+    this.filteredTurmas.sort((a, b) => {
+      let value1 = a[this.sortField];
+      let value2 = b[this.sortField];
+      let result = null;
+
+      if (value1 == null && value2 != null)
+        result = -1;
+      else if (value1 != null && value2 == null)
+        result = 1;
+      else if (value1 == null && value2 == null)
+        result = 0;
+      else if (typeof value1 === 'string' && typeof value2 === 'string')
+        result = value1.localeCompare(value2);
+      else
+        result = (value1 < value2 ? -1 : (value1 > value2 ? 1 : 0));
+
+      return (this.sortOrder * result);
+    });
   }
 }
